HTTP filters refactoring
