+++ http-client/src/main/java/io/micronaut/http/client/netty/DefaultHttpClient.java
@@ -987,7 +987,7 @@ public class DefaultHttpClient implements
 
         // apply filters
         streamResponsePublisher = Flux.from(
-                applyFilterToResponsePublisher(parentRequest, request, requestURI, requestWrapper, streamResponsePublisher)
+                applyFilterToResponsePublisher(parentRequest, request, requestURI, streamResponsePublisher)
         );
 
         return streamResponsePublisher;
@@ -1018,7 +1018,6 @@ public class DefaultHttpClient implements
                                     request,
                                     requestWrapper.get(),
                                     requestURI,
-                                    requestWrapper,
                                     (Publisher) proxyResponsePublisher
                             )
                     );
@@ -1118,7 +1117,6 @@ public class DefaultHttpClient implements
             parentRequest,
             request,
             requestURI,
-            requestWrapper,
             responsePublisher
         );
         Flux<io.micronaut.http.HttpResponse<O>> finalReactiveSequence = Flux.from(finalPublisher);
@@ -1204,16 +1202,14 @@ public class DefaultHttpClient implements
             io.micronaut.http.HttpRequest<?> parentRequest,
             io.micronaut.http.HttpRequest<I> request,
             URI requestURI,
-            AtomicReference<MutableHttpRequest<?>> requestWrapper,
             Publisher<R> responsePublisher) {
 
-        if (!(request instanceof MutableHttpRequest mutRequest)) {
+        if (!(request instanceof MutableHttpRequest<?> mutRequest)) {
             return responsePublisher;
         }
 
         mutRequest.uri(requestURI);
-        if (informationalServiceId != null &&
-            !mutRequest.getAttribute(HttpAttributes.SERVICE_ID).isPresent()) {
+        if (informationalServiceId != null && mutRequest.getAttribute(HttpAttributes.SERVICE_ID).isEmpty()) {
 
             mutRequest.setAttribute(HttpAttributes.SERVICE_ID, informationalServiceId);
         }
@@ -1223,12 +1219,12 @@ public class DefaultHttpClient implements
         if (parentRequest != null) {
             // todo: migrate to new filter
             filters.add(
-                new GenericHttpFilter.AroundLegacy(new ClientServerContextFilter(parentRequest),
-                new FilterOrder.Fixed(Ordered.HIGHEST_PRECEDENCE)));
+                GenericHttpFilter.createLegacyFilter(new ClientServerContextFilter(parentRequest), new FilterOrder.Fixed(Ordered.HIGHEST_PRECEDENCE))
+            );
         }
 
         FilterRunner.sortReverse(filters);
-        filters.add(new GenericHttpFilter.TerminalReactive(responsePublisher));
+        filters.add(GenericHttpFilter.terminalReactive(responsePublisher));
 
         FilterRunner runner = new FilterRunner(filters);
         Mono<R> responseMono = Mono.from(ReactiveExecutionFlow.fromFlow((ExecutionFlow<R>) runner.run(request)).toPublisher());
